rename remove_noncands clean up
