+++ src/votekit/elections/election_types/ranking/alaska.py
@@ -1,7 +1,8 @@
 from .abstract_ranking import RankingElection
 from ....pref_profile import PreferenceProfile
 from ...election_state import ElectionState
-from ....utils import first_place_votes, remove_cand
+from ....cleaning import remove_cand
+from ....utils import first_place_votes
 from ..ranking import Plurality, STV
 from ...transfers import fractional_transfer
 from ....ballot import Ballot
